% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_formats.R
\name{append_to_log}
\alias{append_to_log}
\title{Append metadata to log file}
\usage{
append_to_log(metadata, log_file)
}
\arguments{
\item{metadata}{Metadata object from create_metadata()}

\item{log_file}{Path to log file}
}
\value{
TRUE if successful, FALSE otherwise
}
\description{
Append metadata to log file
}
