% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_system_capabilities}
\alias{check_system_capabilities}
\title{Check system capabilities for metadata storage}
\usage{
check_system_capabilities()
}
\value{
List of capabilities by platform
}
\description{
Check system capabilities for metadata storage
}
