% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{command_exists}
\alias{command_exists}
\title{Helper function to check if command exists on system}
\usage{
command_exists(cmd)
}
\arguments{
\item{cmd}{Command to check}
}
\value{
TRUE if command exists, FALSE otherwise
}
\description{
Helper function to check if command exists on system
}
