% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_formats.R
\name{create_metadata}
\alias{create_metadata}
\title{Create comprehensive metadata object}
\usage{
create_metadata(x, file, script_info)
}
\arguments{
\item{x}{Data object being exported}

\item{file}{File path where data is saved}

\item{script_info}{Script information from get_script_info()}
}
\value{
List containing comprehensive metadata
}
\description{
Create comprehensive metadata object
}
