% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{.exportr_options}
\alias{.exportr_options}
\title{Package configuration functions}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.exportr_options
}
\description{
Set and get package-wide options for exportr
}
\keyword{datasets}
