% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_exportr_options}
\alias{get_exportr_options}
\title{Get exportr package options}
\usage{
get_exportr_options(option_name = NULL)
}
\arguments{
\item{option_name}{Name of option to get, or NULL for all options}
}
\value{
Option value or list of all options
}
\description{
Get exportr package options
}
