% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_detection.R
\name{get_script_info}
\alias{get_script_info}
\title{Intelligent script detection across environments}
\usage{
get_script_info()
}
\value{
A list with script information including:
\itemize{
  \item script_name: Name of the script file
  \item script_path: Full path to the script file (PRIORITY 1)
  \item working_directory: Current working directory
  \item user: Username
  \item execution_time: Timestamp when function was called
  \item detection_method: Method used to detect the script
  \item r_version: R version string
  \item platform: Platform information
}
}
\description{
This function attempts to detect the path of the R script that is currently
being executed using multiple methods across different environments.
}
\details{
Detection methods tried in order:
\enumerate{
  \item RStudio API: \code{rstudioapi::getActiveDocumentContext()$path}
  \item Command line: Parse \code{commandArgs()} for \code{--file=} argument
  \item Call stack analysis: Search sys.frames() for source references
  \item Environment variables: Check for sourcing info
}
}
\examples{
\dontrun{
# Get current script information
script_info <- get_script_info()
print(script_info$script_path)
print(script_info$detection_method)
}

}
