% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_linux_metadata}
\alias{has_linux_metadata}
\title{Check for Linux extended attributes}
\usage{
has_linux_metadata(file)
}
\arguments{
\item{file}{File path}
}
\value{
TRUE if metadata found, FALSE otherwise
}
\description{
Check for Linux extended attributes
}
