% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_macos_metadata}
\alias{has_macos_metadata}
\title{Check for macOS extended attributes}
\usage{
has_macos_metadata(file)
}
\arguments{
\item{file}{File path}
}
\value{
TRUE if metadata found, FALSE otherwise
}
\description{
Check for macOS extended attributes
}
