% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_windows_metadata}
\alias{has_windows_metadata}
\title{Check for Windows alternate data streams}
\usage{
has_windows_metadata(file)
}
\arguments{
\item{file}{File path}
}
\value{
TRUE if metadata found, FALSE otherwise
}
\description{
Check for Windows alternate data streams
}
