% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_linux_metadata}
\alias{read_linux_metadata}
\title{Read Linux extended attributes}
\usage{
read_linux_metadata(file_path)
}
\arguments{
\item{file_path}{File path}
}
\value{
Metadata list
}
\description{
Read Linux extended attributes
}
