% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_macos_metadata}
\alias{read_macos_metadata}
\title{Read macOS extended attributes}
\usage{
read_macos_metadata(file_path)
}
\arguments{
\item{file_path}{File path}
}
\value{
Metadata list
}
\description{
Read macOS extended attributes
}
