% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_sidecar_metadata}
\alias{read_sidecar_metadata}
\title{Read sidecar metadata file}
\usage{
read_sidecar_metadata(sidecar_path)
}
\arguments{
\item{sidecar_path}{Path to sidecar file}
}
\value{
Metadata list
}
\description{
Read sidecar metadata file
}
