% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_windows_metadata}
\alias{read_windows_metadata}
\title{Read Windows alternate data streams}
\usage{
read_windows_metadata(file_path)
}
\arguments{
\item{file_path}{File path}
}
\value{
Metadata list
}
\description{
Read Windows alternate data streams
}
