% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_exportr_options}
\alias{set_exportr_options}
\title{Set exportr package options}
\usage{
set_exportr_options(
  default_metadata_mode = "attribute",
  verbose = TRUE,
  fallback_to_sidecar = TRUE
)
}
\arguments{
\item{default_metadata_mode}{Default metadata mode for all functions}

\item{verbose}{Whether to print verbose output}

\item{fallback_to_sidecar}{Whether to fallback to sidecar files if attributes fail}
}
\description{
Set exportr package options
}
