% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_detection.R
\name{test_script_detection}
\alias{test_script_detection}
\title{Helper function to validate script detection}
\usage{
test_script_detection()
}
\value{
TRUE if script detection is working, FALSE otherwise
}
\description{
Helper function to validate script detection
}
