% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwrite_with_metadata.R
\name{write_comment_metadata}
\alias{write_comment_metadata}
\title{Write metadata as comments to CSV file (discouraged method)}
\usage{
write_comment_metadata(metadata, file)
}
\arguments{
\item{metadata}{Metadata object}

\item{file}{File path}
}
\value{
TRUE if successful, FALSE otherwise
}
\description{
Write metadata as comments to CSV file (discouraged method)
}
