% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_formats.R
\name{write_file_attributes}
\alias{write_file_attributes}
\title{Write metadata to file system properties/extended attributes}
\usage{
write_file_attributes(metadata, file)
}
\arguments{
\item{metadata}{Metadata object from create_metadata()}

\item{file}{File path}
}
\value{
TRUE if successful, FALSE otherwise
}
\description{
Write metadata to file system properties/extended attributes
}
