% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_formats.R
\name{write_macos_attributes}
\alias{write_macos_attributes}
\title{Write metadata using macOS extended attributes}
\usage{
write_macos_attributes(metadata, file)
}
\arguments{
\item{metadata}{Metadata object}

\item{file}{File path}
}
\value{
TRUE if successful, FALSE otherwise
}
\description{
Write metadata using macOS extended attributes
}
