% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_formats.R
\name{write_metadata_file}
\alias{write_metadata_file}
\title{Write metadata to separate file}
\usage{
write_metadata_file(metadata, file, format = "json")
}
\arguments{
\item{metadata}{Metadata object from create_metadata()}

\item{file}{Original file path}

\item{format}{Format for metadata file ("json", "yaml", "txt")}
}
\value{
TRUE if successful, FALSE otherwise
}
\description{
Write metadata to separate file
}
