% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_formats.R
\name{write_sidecar_metadata}
\alias{write_sidecar_metadata}
\title{Write metadata to hidden sidecar file}
\usage{
write_sidecar_metadata(metadata, file)
}
\arguments{
\item{metadata}{Metadata object from create_metadata()}

\item{file}{File path}
}
\value{
TRUE if successful, FALSE otherwise
}
\description{
Write metadata to hidden sidecar file
}
