% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_formats.R
\name{write_windows_attributes}
\alias{write_windows_attributes}
\title{Write metadata using Windows alternate data streams}
\usage{
write_windows_attributes(metadata, file)
}
\arguments{
\item{metadata}{Metadata object}

\item{file}{File path}
}
\value{
TRUE if successful, FALSE otherwise
}
\description{
Write metadata using Windows alternate data streams
}
